+++ src/server/index.ts
@@ -18,6 +18,7 @@ export { LogContextService } from './services/log-context.service'
 // Destinations
 export { DefaultStdoutDestination } from './destinations/default-stdout.destination'
 export { PrettyDevDestination } from './destinations/pretty-dev.destination'
+export type { PrettyViewOptions } from './destinations/pretty-dev.destination'
 
 // HTTP integration (interceptor, filter, middleware)
 export { HttpExceptionFilter } from './filters/http-exception.filter'